requirements.txt
fastapi==0.135.1
uvicorn[standard]
sqlalchemy==2.0.48
jinja2==3.1.6
python-multipart==0.0.22
pydantic==2.12.5
bcrypt==5.0.0
PyJWT==2.12.1
itsdangerous==2.2.0

torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
numpy==2.4.3
scipy==1.17.1
scikit-learn==1.8.0
optuna==4.7.0
h5py==3.16.0

matplotlib==3.10.8
seaborn==0.13.2
pandas==3.0.1
jupyter
ipykernel
